First steps towards message compatibility support
Added a multi vibrator support message: VibrateCmd
Added a client MessageVersion to the RequestServerInfo message

ButtplugMessage base class now contains message versioning boilerplate: this allows tracking of when messages were added and contains references to their legacy counterparts. The aim is that when dealing with an older client, we downgrade the messages to the highest common level.